// web/src/api/models/GithubPRData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GithubPRData
 */
export interface GithubPRData {
    /**
     * 
     * @type {string}
     * @memberof GithubPRData
     */
    author?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubPRData
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubPRData
     */
    htmlUrl?: string;
    /**
     * 
     * @type {number}
     * @memberof GithubPRData
     */
    number?: number;
    /**
     * 
     * @type {string}
     * @memberof GithubPRData
     */
    repo?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubPRData
     */
    title?: string;
}

/**
 * Check if a given object implements the GithubPRData interface.
 */
export function instanceOfGithubPRData(value: object): value is GithubPRData {
    return true;
}

export function GithubPRDataFromJSON(json: any): GithubPRData {
    return GithubPRDataFromJSONTyped(json, false);
}

export function GithubPRDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): GithubPRData {
    if (json == null) {
        return json;
    }
    return {
        
        'author': json['author'] == null ? undefined : json['author'],
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'htmlUrl': json['html_url'] == null ? undefined : json['html_url'],
        'number': json['number'] == null ? undefined : json['number'],
        'repo': json['repo'] == null ? undefined : json['repo'],
        'title': json['title'] == null ? undefined : json['title'],
    };
}

export function GithubPRDataToJSON(json: any): GithubPRData {
    return GithubPRDataToJSONTyped(json, false);
}

export function GithubPRDataToJSONTyped(value?: GithubPRData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'author': value['author'],
        'created_at': value['createdAt'],
        'html_url': value['htmlUrl'],
        'number': value['number'],
        'repo': value['repo'],
        'title': value['title'],
    };
}