// web/src/api/models/GithubPRsWidgetData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GithubPRData } from './GithubPRData';
import {
    GithubPRDataFromJSON,
    GithubPRDataFromJSONTyped,
    GithubPRDataToJSON,
    GithubPRDataToJSONTyped,
} from './GithubPRData';

/**
 * 
 * @export
 * @interface GithubPRsWidgetData
 */
export interface GithubPRsWidgetData {
    /**
     * 
     * @type {Array<GithubPRData>}
     * @memberof GithubPRsWidgetData
     */
    assigned?: Array<GithubPRData>;
    /**
     * 
     * @type {Array<GithubPRData>}
     * @memberof GithubPRsWidgetData
     */
    authored?: Array<GithubPRData>;
    /**
     * 
     * @type {Array<GithubPRData>}
     * @memberof GithubPRsWidgetData
     */
    mentioned?: Array<GithubPRData>;
    /**
     * 
     * @type {Array<GithubPRData>}
     * @memberof GithubPRsWidgetData
     */
    reviewRequested?: Array<GithubPRData>;
}

/**
 * Check if a given object implements the GithubPRsWidgetData interface.
 */
export function instanceOfGithubPRsWidgetData(value: object): value is GithubPRsWidgetData {
    return true;
}

export function GithubPRsWidgetDataFromJSON(json: any): GithubPRsWidgetData {
    return GithubPRsWidgetDataFromJSONTyped(json, false);
}

export function GithubPRsWidgetDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): GithubPRsWidgetData {
    if (json == null) {
        return json;
    }
    return {
        
        'assigned': json['assigned'] == null ? undefined : ((json['assigned'] as Array<any>).map(GithubPRDataFromJSON)),
        'authored': json['authored'] == null ? undefined : ((json['authored'] as Array<any>).map(GithubPRDataFromJSON)),
        'mentioned': json['mentioned'] == null ? undefined : ((json['mentioned'] as Array<any>).map(GithubPRDataFromJSON)),
        'reviewRequested': json['review_requested'] == null ? undefined : ((json['review_requested'] as Array<any>).map(GithubPRDataFromJSON)),
    };
}

export function GithubPRsWidgetDataToJSON(json: any): GithubPRsWidgetData {
    return GithubPRsWidgetDataToJSONTyped(json, false);
}

export function GithubPRsWidgetDataToJSONTyped(value?: GithubPRsWidgetData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'assigned': value['assigned'] == null ? undefined : ((value['assigned'] as Array<any>).map(GithubPRDataToJSON)),
        'authored': value['authored'] == null ? undefined : ((value['authored'] as Array<any>).map(GithubPRDataToJSON)),
        'mentioned': value['mentioned'] == null ? undefined : ((value['mentioned'] as Array<any>).map(GithubPRDataToJSON)),
        'review_requested': value['reviewRequested'] == null ? undefined : ((value['reviewRequested'] as Array<any>).map(GithubPRDataToJSON)),
    };
}